bugs fixed

+++ backend/django_app/project/lawyer/models.py
@@ -9,6 +9,6 @@ class LawyerProfile(models.Model):
         User,
         on_delete=models.CASCADE,
         related_name="lawyer_profile")
-    fees = models.DecimalField(max_digits=10, decimal_places=2)
+    fees = models.IntegerField()
     experience = models.PositiveIntegerField()
     speciality = models.CharField(max_length=50)
